+++ includes/service-providers/mailchimp/class-newspack-newsletters-mailchimp.php
@@ -508,7 +508,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 
 					$groups = array_map(
 						function ( $group ) use ( $list ) {
-							$group['id']   = $this->create_group_or_tag_list_id( $group['id'], $list['id'] );
+							$group['id']   = Subscription_List::mailchimp_generate_public_id( $group['id'], $list['id'] );
 							$group['type'] = 'mailchimp-group';
 							return $group;
 						},
@@ -518,7 +518,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 				}
 
 				foreach ( $all_tags as $tag ) {
-					$tag['id']   = $this->create_group_or_tag_list_id( $tag['id'], $list['id'], 'tag' );
+					$tag['id']   = Subscription_List::mailchimp_generate_public_id( $tag['id'], $list['id'], 'tag' );
 					$tag['type'] = 'mailchimp-tag';
 					$lists[]     = $tag;
 				}
@@ -566,7 +566,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 			foreach ( $categories['categories'] as &$category ) {
 				if ( ! empty( $category['interests']['interests'] ) ) {
 					foreach ( $category['interests']['interests'] as &$interest ) {
-						$local_id = $this->create_group_or_tag_list_id( $interest['id'], $list_id );
+						$local_id = Subscription_List::mailchimp_generate_public_id( $interest['id'], $list_id );
 						if ( isset( $configured_lists[ $local_id ]['name'] ) ) {
 							$interest['local_name'] = $configured_lists[ $local_id ]['name'];
 						}
@@ -598,7 +598,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 		$configured_lists = Newspack_Newsletters_Subscription::get_lists_config();
 		if ( ! empty( $configured_lists ) ) {
 			foreach ( $tags as &$tag ) {
-				$local_id = $this->create_group_or_tag_list_id( $tag['id'], $list_id, 'tag' );
+				$local_id = Subscription_List::mailchimp_generate_public_id( $tag['id'], $list_id, 'tag' );
 				if ( isset( $configured_lists[ $local_id ]['name'] ) ) {
 					$tag['local_name'] = $configured_lists[ $local_id ]['name'];
 				}
@@ -1154,64 +1154,257 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 	}
 
 	/**
-	 * Add contact to a list with multiple groups and/or tags.
+	 * Upserts a contact to the ESP using the provider specific methods.
 	 *
-	 * @param array $contact The contact, as for the add_contact method.
-	 * @param array $lists List IDs to add the contact to.
+	 * Here we group all groups and tags by Audience, so we can add many lists at once.
+	 *
+	 * @param array               $contact The contact, as for the add_contact method.
+	 * @param Subscription_List[] $lists The lists.
+	 * @return array|WP_Error Contact data if it was added, or error otherwise.
 	 */
-	public function add_contact_with_groups_and_tags( $contact, $lists ) {
-		$results = [];
-		$by_list = [];
-		foreach ( $lists as $list_id ) {
-			$group_or_tag_list = $this->maybe_extract_group_or_tag_list( $list_id );
-			if ( $group_or_tag_list && isset( $group_or_tag_list['type'] ) ) {
-				$list_id = $group_or_tag_list['list_id'];
-				if ( 'group' === $group_or_tag_list['type'] && ! isset( $contact['interests'] ) ) {
-					$contact['interests'] = [];
-				}
-				if ( isset( $by_list[ $list_id ] ) ) {
-					$by_list[ $list_id ][] = $group_or_tag_list;
-				} else {
-					$by_list[ $list_id ] = [ $group_or_tag_list ];
-				}
-			} else {
-				// It might be a local list – the list id has to be extracted from the DB.
-				$local_list = Subscription_List::from_form_id( $list_id );
-				if ( $local_list && $local_list->is_configured_for_provider( $this->service ) ) {
-					$list_settings = $local_list->get_provider_settings( $this->service );
-					if ( $list_settings !== null ) {
-						$list_id = $list_settings['list'];
-						$sublist = [
-							'id'      => $list_settings['tag_id'],
-							'list_id' => $list_id,
-							'type'    => 'tag',
-						];
-						if ( isset( $by_list[ $list_id ] ) ) {
-							$by_list[ $list_id ][] = $sublist;
-						} else {
-							$by_list[ $list_id ] = [ $sublist ];
-						}
-					}
-				}
-			}
-			if ( ! isset( $by_list[ $list_id ] ) ) {
-				// If this is not a group-or-tag list, nor a local list, treat the list ID as a regular list.
-				$by_list[ $list_id ] = [];
-			}
-		}
-		if ( empty( $by_list ) ) {
+	public function upsert_contact( $contact, $lists ) {
+
+		$prepared_lists = $this->prepare_lists_to_add_contact( $lists );
+
+		if ( empty( $prepared_lists ) ) {
 			return new WP_Error( 'No lists found.' );
 		}
 
-		foreach ( $by_list as $list_id => $sublists ) {
-			$result = $this->add_contact( $contact, $list_id, $sublists );
+		foreach ( $prepared_lists as $audience_id => $sublists ) {
+			$result = $this->add_contact( $contact, $audience_id, $sublists['tags'], $sublists['interests'] );
 			if ( is_wp_error( $result ) ) {
 				return $result;
 			}
-
-			$results[] = $result;
 		}
-		return $results;
+
+		// on success, return the last resutl.
+		return $result;
+	}
+
+	/**
+	 * Loops through a list of Subscription_Lists objects and group them to be sent to the `add_contact` method with tags and interests attached.
+	 *
+	 * @param Subscription_List[] $lists The lists.
+	 * @return array The lists array where the keys are the Audience IDs and the values are an array with 'tags' and 'interests' keys.
+	 */
+	private function prepare_lists_to_add_contact( $lists ) {
+
+		$result = [];
+
+		foreach ( $lists as $list ) {
+
+			$audience_id = $list->mailchimp_get_audience_id();
+
+			if ( ! isset( $result[ $audience_id ] ) ) {
+				$result[ $audience_id ] = [
+					'tags'      => [],
+					'interests' => [],
+				];
+			}
+
+			if ( 'tag' === $list->mailchimp_get_sublist_type() ) {
+				// Mailchimp API expects the tag name and not ID.
+				$result[ $audience_id ]['tags'][] = $list->get_remote_name();
+			} elseif ( 'group' === $list->mailchimp_get_sublist_type() ) {
+				// Local lists are included here.
+				$result[ $audience_id ]['interests'][ $list->mailchimp_get_sublist_id() ] = true;
+			}
+		}
+
+		return $result;
+	}
+
+	/**
+	 * Get merge field type.
+	 *
+	 * @param mixed $value Value to check.
+	 *
+	 * @return string Merge field type.
+	 */
+	private function get_merge_field_type( $value ) {
+		if ( is_numeric( $value ) ) {
+			return 'number';
+		}
+		if ( is_bool( $value ) ) {
+			return 'boolean';
+		}
+		return 'text';
+	}
+
+	/**
+	 * Given a contact metadata array, build the `merge_fields` array to be sent to Mailchimp
+	 * by sarching for existing merge fields and creating new ones as needed.
+	 *
+	 * @param string $audience_id Audience ID.
+	 * @param array  $contact     The contact.
+	 *
+	 * @return array Merge fields.
+	 */
+	private function prepare_merge_fields( $audience_id, $contact ) {
+		$mc           = new Mailchimp( $this->api_key() );
+		$merge_fields = [];
+		$data         = $contact['metadata'];
+
+		// Strip arrays and statuses.
+		$data = array_filter(
+			$data,
+			function( $value, $key ) {
+				return ! is_array( $value ) && 'status' !== $key && 'status_if_new' !== $key;
+			},
+			ARRAY_FILTER_USE_BOTH
+		);
+
+		// Get and match existing merge fields.
+		try {
+			$existing_fields = $mc->get(
+				"lists/$audience_id/merge-fields",
+				[
+					'count' => 1000,
+				],
+				60
+			)['merge_fields'];
+		} catch ( \Exception $e ) {
+			do_action(
+				'newspack_log',
+				'newspack_mailchimp_prepare_merge_fields',
+				sprintf( 'Error getting merge fields: %s', $e->getMessage() ),
+				[
+					'type'       => 'error',
+					'data'       => [
+						'audience_id' => $audience_id,
+						'error'       => $e->getMessage(),
+					],
+					'user_email' => $contact['email'],
+					'file'       => 'newspack_mailchimp',
+				]
+			);
+			return [];
+		}
+		if ( empty( $existing_fields ) ) {
+			$existing_fields = [];
+		}
+
+		usort(
+			$existing_fields,
+			function( $a, $b ) {
+				return $a['merge_id'] - $b['merge_id'];
+			}
+		);
+
+		$list_merge_fields = [];
+
+		// Handle duplicate fields.
+		foreach ( $existing_fields as $field ) {
+			if ( ! isset( $list_merge_fields[ $field['name'] ] ) ) {
+				$list_merge_fields[ $field['name'] ] = $field['tag'];
+			} else {
+				do_action(
+					'newspack_log',
+					'newspack_mailchimp_prepare_merge_fields',
+					sprintf( 'Duplicate merge field %1$s found with tag %2$s.', $field['name'], $field['tag'] ),
+					[
+						'type'       => 'error',
+						'data'       => [
+							'audience_id' => $audience_id,
+							'field'       => $field,
+						],
+						'user_email' => $contact['email'],
+						'file'       => 'newspack_mailchimp',
+					]
+				);
+			}
+		}
+
+		foreach ( $data as $field_name => $field_value ) {
+			// If field already exists, add it to the payload.
+			if ( isset( $list_merge_fields[ $field_name ] ) ) {
+				$merge_fields[ $list_merge_fields[ $field_name ] ] = $data[ $field_name ];
+				unset( $data[ $field_name ] );
+			}
+		}
+
+		// Create remaining fields.
+		$remaining_fields = array_keys( $data );
+		foreach ( $remaining_fields as $field_name ) {
+			$field_data = [
+				'name' => $field_name,
+				'type' => $this->get_merge_field_type( $data[ $field_name ] ),
+			];
+			$created_field = $mc->post( "lists/$audience_id/merge-fields", $field_data );
+			if ( empty( $created_field['merge_id'] ) ) {
+				$message = sprintf(
+					// Translators: %1$s is the merge field key, %2$s is the error message.
+					__( 'Failed to create merge field %1$s. Error response: %2$s', 'newspack-newsletters' ),
+					$field_name,
+					$created_field['detail'] ?? 'Unknown error'
+				);
+			} else {
+				$message = sprintf(
+					// Translators: %1$s is the merge field key, %2$s is the merge field tag.
+					__( 'Created merge field %1$s with tag %2$s.', 'newspack-newsletters' ),
+					$field_name,
+					$created_field['tag']
+				);
+			}
+			do_action(
+				'newspack_log',
+				'newspack_mailchimp_prepare_merge_fields',
+				$message,
+				[
+					'type'       => empty( $created_field['merge_id'] ) ? 'error' : 'debug',
+					'data'       => [
+						'audience_id'   => $audience_id,
+						'field_data'    => $field_data,
+						'created_field' => $created_field,
+					],
+					'user_email' => $contact['email'],
+					'file'       => 'newspack_mailchimp',
+				]
+			);
+			// Add the field to the merge fields array if it was created.
+			if ( ! empty( $created_field['merge_id'] ) ) {
+				$merge_fields[ $created_field['tag'] ] = $data[ $field_name ];
+			}
+		}
+
+		return $merge_fields;
+	}
+
+	/**
+	 * Gets the status and/or status_if_new keys based on the contact data.
+	 *
+	 * @param array  $contact      {
+	 *    Contact data.
+	 *
+	 *    @type string   $email    Contact email address.
+	 *    @type string   $name     Contact name. Optional.
+	 *    @type string[] $metadata Contact additional metadata. Optional.
+	 * }
+	 * @param string $list_id List (Audience) to add the contact to, if any.
+	 *
+	 * @return array The status and/or status_if_new keys to be added to the payload
+	 */
+	private function get_status_for_payload( $contact, $list_id = null ) {
+		$return = [];
+		if ( ! empty( $contact['metadata']['status_if_new'] ) ) {
+			$return['status_if_new'] = $contact['metadata']['status_if_new'];
+		}
+
+		if ( ! empty( $contact['metadata']['status'] ) ) {
+			$return['status'] = $contact['metadata']['status'];
+		}
+
+		// Check if the contact has unsubscribed before. Mailchimp requires a double opt-in to resubscribe, so we set the status to 'pending'.
+		if ( $list_id && ! empty( $contact['existing_contact_data']['lists'][ $list_id ]['status'] ) && 'unsubscribed' === $contact['existing_contact_data']['lists'][ $list_id ]['status'] ) {
+			$return['status'] = 'pending';
+		}
+
+		// If we're subscribing the contact to a newsletter, they should have some status
+		// because 'non-subscriber' status can't receive newsletters.
+		if ( empty( $return['status'] ) && empty( $return['status_if_new'] ) ) {
+			$return['status'] = 'subscribed';
+		}
+		return $return;
 	}
 
 	/**
@@ -1224,12 +1417,13 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 	 *    @type string   $name     Contact name. Optional.
 	 *    @type string[] $metadata Contact additional metadata. Optional.
 	 * }
-	 * @param string $list_id      List to add the contact to.
-	 * @param string $sublists     An array of groups and/or tags in the list to add the contact to.
+	 * @param string $list_id   List (Audience) to add the contact to.
+	 * @param array  $tags      An array of tag names to be added to the contact. Tags are always appended.
+	 * @param array  $interests An array of interests as expected by the API, where the key is the interest ID and the value is a bool (add or remove).
 	 *
 	 * @return array|WP_Error Contact data if it was added, or error otherwise.
 	 */
-	public function add_contact( $contact, $list_id = false, $sublists = [] ) {
+	public function add_contact( $contact, $list_id = false, $tags = [], $interests = [] ) {
 		if ( false === $list_id ) {
 			return new WP_Error( 'newspack_newsletters_mailchimp_list_id', __( 'Missing list id.' ) );
 		}
@@ -1237,140 +1431,52 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 
 		// If contact was added in this execution, we can return the previous
 		// result and bail.
-		$cache_key = $list_id . $email_address . wp_json_encode( $sublists );
+		$cache_key = md5( $list_id . $email_address . wp_json_encode( $tags ) . wp_json_encode( $interests ) );
 		if ( ! empty( self::$contacts_added[ $cache_key ] ) ) {
 			return self::$contacts_added[ $cache_key ];
 		}
 
-		$new_contact_status = 'subscribed';
-		if ( isset( $contact['metadata'] ) && ! empty( $contact['metadata']['status'] ) ) {
-			$new_contact_status = $contact['metadata']['status'];
-			unset( $contact['metadata']['status'] );
+		$update_payload = [ 'email_address' => $email_address ];
+
+		$update_payload = array_merge(
+			$update_payload,
+			$this->get_status_for_payload( $contact, $list_id )
+		);
+
+		// Parse full name into first + last.
+		if ( isset( $contact['name'] ) ) {
+			$name_fragments = explode( ' ', $contact['name'], 2 );
+			$contact['metadata']['First Name'] = $name_fragments[0];
+			if ( isset( $name_fragments[1] ) ) {
+				$contact['metadata']['Last Name'] = $name_fragments[1];
+			}
 		}
+
 		try {
-			$mc             = new Mailchimp( $this->api_key() );
-			$update_payload = [ 'email_address' => $email_address ];
+
+			$mc = new Mailchimp( $this->api_key() );
 
 			if ( isset( $contact['metadata'] ) && is_array( $contact['metadata'] ) && ! empty( $contact['metadata'] ) ) {
-				$update_payload['merge_fields'] = [];
-
-				$merge_fields_res = $mc->get( "lists/$list_id/merge-fields", [ 'count' => 1000 ] );
-				if ( ! isset( $merge_fields_res['merge_fields'] ) ) {
-					return new \WP_Error(
-						'newspack_newsletters_mailchimp_add_contact_failed',
-						sprintf(
-							// Translators: %1$s is the error message.
-							__( 'Error getting merge fields: %1$s', 'newspack-newsletters' ),
-							$merge_fields_res['detail'] ?? __( 'Unable to fetch merge fields.' )
-						)
-					);
+				$merge_fields = $this->prepare_merge_fields( $list_id, $contact );
+				if ( ! empty( $merge_fields ) ) {
+					$update_payload['merge_fields'] = $merge_fields;
 				}
-				$existing_merge_fields = $merge_fields_res['merge_fields'];
-				usort(
-					$existing_merge_fields,
-					function ( $a, $b ) {
-						return $a['merge_id'] - $b['merge_id'];
-					}
-				);
-
-				$list_merge_fields = [];
-
-				// Handle duplicate fields.
-				foreach ( $existing_merge_fields as $key => $field ) {
-					if ( ! isset( $list_merge_fields[ $field['name'] ] ) ) {
-						$list_merge_fields[ $field['name'] ] = $field['tag'];
-					} else {
-						Newspack_Newsletters_Logger::log(
-							sprintf(
-								// Translators: %1$s is the merge field name, %2$s is the unique tag.
-								__( 'Warning: Duplicate merge field %1$s found with tag %2$s.', 'newspack-newsletters' ),
-								$field['name'],
-								$field['tag']
-							)
-						);
-					}
-				}
-
-				foreach ( $contact['metadata'] as $key => $value ) {
-					if ( isset( $list_merge_fields[ $key ] ) ) {
-						$update_payload['merge_fields'][ $list_merge_fields[ $key ] ] = (string) $value;
-					} else {
-						$created_merge_field = $mc->post(
-							"lists/$list_id/merge-fields",
-							[
-								'name' => $key,
-								'type' => 'text',
-							]
-						);
-						if ( isset( $created_merge_field['status'] ) && '4' === substr( $created_merge_field['status'], 0, 1 ) ) {
-							Newspack_Newsletters_Logger::log(
-								sprintf(
-									// Translators: %1$s is the merge field key, %2$s is the error message.
-									__( 'Failed to create merge field %1$s. Reason: %2$s', 'newspack-newsletters' ),
-									$key,
-									$created_merge_field['detail'] ?? $created_merge_field['title']
-								)
-							);
-							continue;
-						}
-						$update_payload['merge_fields'][ $created_merge_field['tag'] ] = (string) $value;
-					}
-				}
-			}
-
-			$all_tags = [];
-			// If the sublists contain any tags, fetch all tags.
-			// This is needed because MC API expects tag names in the contact update payload, not ids.
-			if ( array_filter(
-				$sublists,
-				function( $sublist ) {
-					return 'tag' === $sublist['type']; }
-			) ) {
-				$all_tags = array_reduce(
-					$mc->get( "lists/$list_id/tag-search" )['tags'],
-					function( $tags, $tag ) {
-						$tags[ $tag['id'] ] = $tag['name'];
-						return $tags;
-					},
-					[]
-				);
 			}
 
 			// Add groups and tags, if any.
-			foreach ( $sublists as $sublist ) {
-				$sublist_id   = $sublist['id'];
-				$sublist_type = $sublist['type'];
-				if ( 'group' === $sublist_type ) {
-					if ( ! isset( $update_payload['interests'] ) ) {
-						$update_payload['interests'] = [];
-					}
-					$update_payload['interests'][ $sublist['id'] ] = true;
-				} elseif ( 'tag' === $sublist_type ) {
-					if ( ! isset( $update_payload['tags'] ) ) {
-						$update_payload['tags'] = [];
-					}
-					if ( isset( $all_tags[ $sublist_id ] ) ) {
-						$update_payload['tags'][] = $all_tags[ $sublist_id ];
-					}
-				}
+			if ( ! empty( $tags ) ) {
+				$update_payload['tags'] = $tags;
+			}
+			if ( ! empty( $interests ) ) {
+				$update_payload['interests'] = $interests;
 			}
 
-			// If we're subscribing the contact to a newsletter, they should have some status
-			// because 'non-subscriber' status can't receive newsletters.
-			if ( ! empty( $list_id ) || ! empty( $sublists ) ) {
-				$update_payload['status_if_new'] = $new_contact_status ?? 'subscribed';
-				$update_payload['status']        = $new_contact_status ?? 'subscribed';
-			}
+			Newspack_Newsletters_Logger::log( 'Mailchimp add_contact PUT payload: ' . wp_json_encode( $update_payload ) );
 
 			// Create or update a list member.
-			$existing_contact = self::get_contact_data( $email_address );
-			if ( is_wp_error( $existing_contact ) ) {
-				$update_payload['status'] = $new_contact_status ?? 'subscribed';
-				$result                   = $mc->post( "lists/$list_id/members", $update_payload );
-			} else {
-				$member_id = $existing_contact['id'];
-				$result    = $mc->put( "lists/$list_id/members/$member_id", $update_payload );
-			}
+			$member_hash = Mailchimp::subscriberHash( $email_address );
+			$result = $mc->put( "lists/$list_id/members/$member_hash", $update_payload );
+
 			if (
 				! $result ||
 				! isset( $result['status'] ) ||
@@ -1448,14 +1554,14 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 		foreach ( $contact['interests'] as $list_id => $interests ) {
 			foreach ( $interests as $group_id => $active ) {
 				if ( $active ) {
-					$groups_lists[] = $this->create_group_or_tag_list_id( $group_id, $list_id );
+					$groups_lists[] = Subscription_List::mailchimp_generate_public_id( $group_id, $list_id );
 				}
 			}
 		}
 		$tags_lists = [];
 		foreach ( $contact['tags'] as $list_id => $tags ) {
 			foreach ( $tags as $tag ) {
-				$tags_lists[] = $this->create_group_or_tag_list_id( $tag['id'], $list_id, 'tag' );
+				$tags_lists[] = Subscription_List::mailchimp_generate_public_id( $tag['id'], $list_id, 'tag' );
 			}
 		}
 		return array_merge( $audience_lists, $groups_lists, $tags_lists );
@@ -1474,7 +1580,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 		$contact = $this->get_contact_data( $email );
 		if ( is_wp_error( $contact ) ) {
 			/** Create contact */
-			$result = Newspack_Newsletters_Subscription::add_contact( [ 'email' => $email ], $lists_to_add );
+			$result = Newspack_Newsletters_Contacts::upsert( [ 'email' => $email ], $lists_to_add );
 			if ( is_wp_error( $result ) ) {
 				return $result;
 			}
@@ -1482,19 +1588,36 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 		}
 		$mc = new Mailchimp( $this->api_key() );
 		try {
+			// Remove lists.
 			foreach ( $lists_to_remove as $list_id ) {
-				$list = $this->maybe_extract_group_or_tag_list( $list_id );
-				if ( $list ) {
-					$this->remove_group_or_tag_from_contact( $email, $list['id'], $list['list_id'], $list['type'] );
+				$list_obj = Subscription_List::from_public_id( $list_id );
+				if ( ! $list_obj ) {
 					continue;
 				}
-				if ( ! isset( $contact['lists'][ $list_id ] ) ) {
-					continue;
+
+				if ( 'group' === $list_obj->mailchimp_get_sublist_type() ) {
+					$this->remove_group_from_contact( $email, $list_obj->mailchimp_get_sublist_id(), $list_obj->mailchimp_get_audience_id() );
+				} elseif ( 'tag' === $list_obj->mailchimp_get_sublist_type() ) {
+					$this->remove_tag_from_contact( $email, $list_obj->mailchimp_get_sublist_id(), $list_obj->mailchimp_get_audience_id() );
 				}
-				$mc->patch( "lists/$list_id/members/" . $contact['lists'][ $list_id ]['contact_id'], [ 'status' => 'unsubscribed' ] );
+
+				// If $list_id is an Audience the contact is a part of, unsubscribe them.
+				if ( isset( $contact['lists'][ $list_id ] ) ) {
+					$mc->patch( "lists/$list_id/members/" . $contact['lists'][ $list_id ]['contact_id'], [ 'status' => 'unsubscribed' ] );
+				}
 			}
+
+			// Add lists.
 			if ( ! empty( $lists_to_add ) ) {
-				$this->add_contact_with_groups_and_tags( [ 'email' => $email ], $lists_to_add );
+				$lists_objects = [];
+				foreach ( $lists_to_add as $list_id ) {
+					$list_obj = Subscription_List::from_public_id( $list_id );
+					if ( ! $list_obj ) {
+						continue;
+					}
+					$lists_objects[] = $list_obj;
+				}
+				$this->upsert_contact( [ 'email' => $email ], $lists_objects );
 			}
 		} catch ( \Exception $e ) {
 			return new \WP_Error(
@@ -1515,12 +1638,18 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 	 */
 	public function get_contact_data( $email, $return_details = false ) {
 		$mc    = new Mailchimp( $this->api_key() );
-		$found = $mc->get(
+		$result  = $mc->get(
 			'search-members',
 			[
 				'query' => $email,
 			]
-		)['exact_matches']['members'];
+		);
+
+		if ( ! isset( $result['exact_matches']['members'] ) ) {
+			return new WP_Error( 'newspack_newsletters_mailchimp_search_members', __( 'Error reaching to search-members endpoint', 'newspack-newsletters' ) );
+		}
+
+		$found = $result['exact_matches']['members'];
 		if ( empty( $found ) ) {
 			return new WP_Error( 'newspack_newsletters_mailchimp_contact_not_found', __( 'Contact not found', 'newspack-newsletters' ) );
 		}
@@ -1603,7 +1732,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 
 			if ( ! empty( $tags[ $list_settings['list'] ] ) ) {
 				if ( in_array( $list_settings['tag_id'], $tags[ $list_settings['list'] ], false ) ) { // phpcs:ignore WordPress.PHP.StrictInArray.FoundNonStrictFalse
-					$ids[] = $list->get_form_id();
+					$ids[] = $list->get_public_id();
 				}
 			}
 		}
@@ -1671,42 +1800,6 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 			'Error sending test email. This campaign cannot be tested:" A From Name must be entered on the Setup step."' => __( 'Error sending test email. Please enter a name and email in the "FROM" section.', 'newspack-newsletters' ),
 		];
 		return isset( $known_errors[ $message ] ) ? $known_errors[ $message ] : $message;
-	}
-
-	/**
-	 * Creates a list ID based on the type, the ID and the list ID
-	 *
-	 * In Mailchimp, we offer both Audiences, Groups, and Tags as Subscription Lists. We modify the group and tag IDs so we can differentiate them from the Audiences IDs.
-	 *
-	 * Also, when working with groups or tags, we need to know the list ID, so we add it to the ID.
-	 *
-	 * @param string $item_id The item ID.
-	 * @param string $list_id The List/Audience ID.
-	 * @param string $type 'group' or 'tag'.
-	 * @return string
-	 */
-	public function create_group_or_tag_list_id( $item_id, $list_id, $type = 'group' ) {
-		return $type . '-' . $item_id . '-' . $list_id;
-	}
-
-	/**
-	 * Extract the group or tag + audience (list) ID from an ID created with create_group_or_tag_list_id
-	 *
-	 * @param string $list_id The list ID.
-	 * @return array|false Array with the group/tag ID and the list ID or false if the ID is not a group or tag list ID.
-	 */
-	public function maybe_extract_group_or_tag_list( $list_id ) {
-		$pattern = '/^(group|tag)-([^-]+)-([^-]+)$/';
-		if ( preg_match( $pattern, $list_id, $matches ) ) {
-			$extracted_ids = [
-				'id'      => $matches[2],
-				'list_id' => $matches[3],
-				'type'    => $matches[1],
-			];
-
-			return $extracted_ids;
-		}
-		return false;
 	}
 
 	/**