+++ tests/test-subscription-attempts.php
@@ -20,6 +20,15 @@ class Subscription_Attempts_Test extends WP_UnitTestCase {
 	}
 
 	/**
+	 * Set up
+	 */
+	public function set_up() {
+		global $wpdb;
+		$table_name = Newspack_Newsletters_Subscription_Attempts::get_table_name();
+		$wpdb->query( "DELETE FROM $table_name" ); // phpcs:ignore
+	}
+
+	/**
 	 * Test if the attempt is added to the custom table when the WP hook is called.
 	 */
 	public function test_subscription_attempts_add_and_update() {
@@ -34,14 +43,14 @@ class Subscription_Attempts_Test extends WP_UnitTestCase {
 		self::assertEquals( implode( ',', $lists ), $result->list_ids );
 
 		$lists_added = [ 'list3', 'list4' ];
-		do_action( 'newspack_newsletters_update_contact_lists', 'some_esp', $contact['email'], $lists_added, [], true );
+		do_action( 'newspack_newsletters_update_contact_lists', 'some_esp', $contact['email'], $lists_added, [], true, 'test' );
 
 		$result = Newspack_Newsletters_Subscription_Attempts::get_by_email( $contact['email'] );
 		$lists_expected = array_merge( $lists, $lists_added );
 		self::assertEquals( implode( ',', $lists_expected ), $result->list_ids );
 
 		$lists_removed = [ 'list1', 'list3' ];
-		do_action( 'newspack_newsletters_update_contact_lists', 'some_esp', $contact['email'], [], $lists_removed, true );
+		do_action( 'newspack_newsletters_update_contact_lists', 'some_esp', $contact['email'], [], $lists_removed, true, 'test' );
 
 		$result = Newspack_Newsletters_Subscription_Attempts::get_by_email( $contact['email'] );
 		$lists_expected = [ 'list2', 'list4' ];